Math.Matrix

Description

This module exports a matrix type as well as some functions to work with it

Types

data Matrix t

a matrix. The constructor is hidden, so it cannot be used directly - use m or mUnsafe instead

type aliases for indexes and width or height

type MatrIndex = (IndexRow, IndexCol)

index to access elements in a matrix. First entry is for the row, second for the column

type IndexRow = Int

type IndexCol = Int

type MatrSize t = Vec Int

size of a matrix. First entry is for the number of rows, second for the number of columns

type Width = Int

type Height = Int

Matrix Pseudo Constructors

m :: MatrSize t -> (MatrIndex -> t) -> Matrix t

creates a matrix from a size and a function

mFromListRow :: Monad m => [[t]] -> m (Matrix t)

creates a matrix from a list of lines. The function fails, if the input is malformed

Getter

mGet :: MatrIndex -> Matrix t -> t

retrieve the element by its index from the matrix

mGetSize :: Matrix e -> (IndexRow, IndexCol)

creates a matrix from a list of lists. If the input is invalid, error is called

create a square matrix

returns a tuple representing the size of the matrix first entry represents the number of rows second entry represents the number of columns

mGetHeight :: Matrix t -> Height

retrieve the height of a matrix, that is the number of lines it consists of

mGetWidth :: Matrix t -> Width

retrieve the width of a matrix, that is the number of columns it consists of

Lists of Indices

mGetAllIndexRow :: Matrix t -> [Height]

mGetAllIndexCol :: Matrix t -> [Width]

mGetAllIndex :: Matrix t -> [(Height, Width)]

extract specific parts from the matrix

mGetSub :: ((IndexRow, IndexCol), MatrSize t) -> Matrix t -> Matrix t

returns a submatrix

mGetRow :: IndexRow -> Matrix b -> [b]

returns a row of the matrix as a list

mGetCol :: IndexCol -> Matrix b -> [b]

returns a column of the matrix as a list

Monadic Getters

mGetWithOrigin :: MatrIndex -> Matrix t -> (t, MatrIndex)

returns an element, packed in the WithOriginMatr Monad

Setter

mSet :: MatrIndex -> t -> Matrix t -> Matrix t

set the element at a specific index inside a matrix

enhanced mapping

mapWithIndex :: (MatrIndex -> a -> b) -> Matrix a -> Matrix b

fmap allows to map a function over a Foldable but the function does not know the position of the element it is applied on. mapWithIndex solves this problem

Special Monads (experimental)

type LogOrigin t = Writer (Log (Origin MatrIndex)) t

This Monad enables you to make calculations based on values that possibly come from a matrix, while logging where the values come from Example:

>>> let test = mUnsafe [[1,2],[3,4]] -- creating a test matrix

>>> runWriter $ mGetWithOrigin (0,1) test
(2,(0,1))

first entry is the value, second is the log of this value (in fact packed into Origin):

>>> :t runWriter $ mGetWithOrigin (0,1) ma
:: (Integer, Log (Origin MatrIndex))

data Origin t

The encodes that the origin might be given:

Constructors

ValO t
origin

NilO
origin not known

type WithLog t = Writer (Log (LogVal t)) t

newtype Log logType

defines a log, which is considered a sequence of any type. if printed it looks like this:

thing0 -> thing1 -> thing2 -> ...

Constructors

Log

Fields

getLog :: [logType]

data LogVal t

Constructors

NilL

ValL t

Fun String