Text.PrettyShow

data types and type classes

class Monoid2D a where

things that can be concatenated in two ways. (|||) is for horizontal, (===) for vertical concatenation

Methods

(|||) :: a -> a -> a

(===) :: a -> a -> a

m2size :: a -> Size Int

m2empty :: a

data RenderMethod src dest

this type represents a method (let us call it a function) to create some data of type t, that fills a frame of some given size

Constructors

RenderMeth

Fields

runRenderMeth :: Size Int -> src -> dest

type SpaceDivide = Size Int -> (Size Int, Size Int)

if there are 2 functions that create a Monoid2D of the same type, the results can be combined generically

type Area a = (Pos a, Size a)

type Size a = Vec a

type Pos a = Vec a

render function combinators

lr :: Monoid2D repr => SpaceDivide -> RenderMethod src repr -> RenderMethod src repr -> RenderMethod (src, src) repr

ud :: Monoid2D repr => SpaceDivide -> RenderMethod src repr -> RenderMethod src repr -> RenderMethod (src, src) repr

vertical :: Monoid2D repr => [RenderMethod src repr] -> RenderMethod [src] repr

horizontal :: Monoid2D repr => [RenderMethod src repr] -> RenderMethod [src] repr

horizontalWith :: Monoid2D repr => (Size Int -> repr) -> [RenderMethod src repr] -> RenderMethod [src] repr

space division functions

horiSpaceDiv :: (Integral b, RealFrac t) => t -> (b, b) -> ((b, b), (b, b))

vertSpaceDiv :: (Integral t1, RealFrac t) => t -> (t1, t1) -> ((t1, t1), (t1, t1))