Text.TextBlock

data types

data TextBlock

represents text in a way that also encodes how it is divided into lines any text t is guaranteed to fit into the rectangle of msize t

nice aliases

type Ellipse = String

type Line = String

pseudo constructors

textBlock :: String -> TextBlock

textBlockTrunc :: (Int, Int) -> String -> TextBlock

textBlockTruncWE :: (Int, Int) -> [Char] -> String -> String -> TextBlock

textBlockAutoNewLineWE :: (Int, Int) -> [Char] -> String -> String -> TextBlock

filledBlock :: String -> Size Int -> TextBlock

force :: Show a => RenderMethod a TextBlock

just forces something into the given size, cut if too big

justBlock :: Show a => RenderMethod a TextBlock

ignores the given size, and renders taking the size needed